Split middleware_test.go into multiple files
